Modified Pacman to use array coordinates for consistency

+++ Game/Character.pde
@@ -2,10 +2,10 @@ interface Character {
   void moveX();
   void moveY();
   void display(float x, float y);
-  float getX();
-  float getY();
-  void setX(float newX);
-  void setY(float newY);
+  int getX();
+  int getY();
+  void setX(int newX);
+  void setY(int newY);
   float getSpeed();;
   void setSpeed(float newSpeed);
